requirements.txt
# MCP-Flow Dependencies

# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
pyyaml>=6.0
requests>=2.31.0

# Web scraping
playwright>=1.40.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Embeddings
sentence-transformers>=2.2.0
faiss-gpu>=1.7.2  # or faiss-cpu if no GPU

# Model training
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
accelerate>=0.25.0
datasets>=2.15.0
bitsandbytes>=0.41.0

# Utilities
tqdm>=4.66.0
coloredlogs>=15.0

# Optional: For DeepSpeed acceleration
# deepspeed>=0.12.0
